+++ react/ProductSummaryDescription.ts
@@ -1,3 +1,0 @@
-import ProductSummaryDescription from './components/ProductSummaryDescription/ProductSummaryDescription'
-
-export default ProductSummaryDescription