+++ react/typings/vtex.render-runtime.d.ts
@@ -1,0 +1,7 @@
+export {}
+
+declare module 'vtex.render-runtime' {
+  export function useTreePath(): {
+    treePath: string
+  }
+}