+++ react/typings/vtex.store-components.d.ts
@@ -1,4 +1,0 @@
-declare module 'vtex.store-components' {
-  const SKUSelector: any
-  const SanitizedHTML: any
-}